+++ CHANGELOG.md
@@ -1,11 +1,21 @@
+# version 0.0.2
+
+- implemented `rremind check` to go through the calendar files and report lines that cannot be interpreted.
+
+- implemented `rremind help`
+
 # version 0.0.1
+
+- Created "remind" clone in Rust for fun and because "Unrecognized command; interpreting as REM" flooded my terminal
 
 - Created "remind" clone in Rust for fun and because "Unrecognized command; interpreting as REM" flooded my terminal
 
 # Unreleased
 
 [ ] Make path to .rem-folder mutable
-[ ] Add formatted output (for terminal)
-[ ] Add 'purge' option
+[ ] Accept date-range input: rremind 0..7 to show appointments of the next 7 days
+[ ] archive dates that are past
+[ ] Make date format more flexible
+[ ] Add nicely formatted output (for terminal)
 [ ] Add .deb package